Removes AnimatedGifView in favor of Emojis, adds ForEach
In order to make the lesson a bit simpler, this commit removes the gif
references and switchs to a simple emoji string for easier image
rendering. Additionally, it makes the VStack View contain a ForEach that
removes a lot of duplicate code and randomizes which cards are face
up/down.